Add ability for deterministic generation
In order to generate the same piece multiple times, we need a way to
make generation deterministic. We do that by seeding all random number
generators with the sum of:
 * the hashCode() of the name of the class that requires random numbers
 * a seed that the user can specify

Above is implemented by the RandomFactory class. If no seed is provided,
all classes that use random numbers work as before - having their random
number generators seeded by the runtime. If the seed is provided, we
seed all random number generators, deterministically, via the sum
described above.

This method assumes that a single invocation of the Generator doesn't
run in parallel, on multiple threads. Separate invocations, in separate
threads, should be supported, though.

Introduce the `seed` command line option in Generator.main().

Adding tests will be done in a separate commit.