Support 256 bit integer seeds for RandomFactory
Seed the RandomFactory with a 256 bit integer. From there, seed every
subsequent Random object with the least significant 64 bits of the
integer, shift 64 bits to the right and then repeat. That way, we use
all the 256 bit randomness.

Additionally, make sure only a single thread uses the RandomFactory,
ensuring deterministic results for the same 256 bit seed.